+++ R/bivariateAMHCopula.R
@@ -1,7 +1,8 @@
 #' Bivariate Ali-Mikhail-Haq Copula
 #'
 #' @description
-#' Computes CDF of the bivariate Ali-Mikhail-Haq copula.
+#' Computes CDF, PDF and simulations of of the bivariate Ali-Mikhail-Haq
+#' copula.
 #'
 #' @details
 #'  The bivariate Ali-Mikhail-Haq copula has CDF :