+++ R/bivariateCACopula.R
@@ -1,7 +1,7 @@
 #' Bivariate Cuadras-Augé Copula
 #'
 #' @description
-#' Computes CDF of the bivariate Cuadras-Augé copula.
+#' Computes CDF and simulations of the bivariate Cuadras-Augé copula.
 #'
 #' @details
 #' The bivariate Cuadras-Augé copula has CDF :