+++ R/bivariateEFGMCopula.R
@@ -1,7 +1,7 @@
 #' Bivariate Eyraud-Farlie-Gumbel-Morgenstern (EFGM) Copula
 #'
 #' @description
-#' Computes CDF of the EFGM copula.
+#' Computes CDF, PDF and simulations of the EFGM copula.
 #'
 #' @details
 #'  The EFGM copula has CDF :