+++ R/bivariateFrankCopula.R
@@ -1,7 +1,7 @@
 #' Bivariate Frank Copula
 #'
 #' @description
-#' Computes CDF of the bivariate Frank copula.
+#' Computes CDF, PDF and simulations of the bivariate Frank copula.
 #'
 #' @details
 #'  The bivariate Frank copula has CDF :