+++ R/bivariateGumbelCopula.R
@@ -1,7 +1,7 @@
 #' Bivariate Gumbel Copula
 #'
 #' @description
-#' Computes CDF of the bivariate Gumbel copula.
+#' Computes CDF, PDF and simulations of the bivariate Gumbel copula.
 #'
 #' @details
 #'  The bivariate Gumbel copula has CDF :