+++ R/frechetCopula.R
@@ -1,7 +1,7 @@
 #' Fréchet Copula
 #'
 #' @description
-#' Computes CDF and PDF of the Fréchet copula.
+#' Computes CDF and simulations of the Fréchet copula.
 #'
 #' @details
 #'  The Fréchet copula has CDF :