+++ R/frechetUpperBoundCopula.R
@@ -1,7 +1,7 @@
 #' Fréchet Upper Bound Copula
 #'
 #' @description
-#' Computes CDF of the Fréchet upper bound copula.
+#' Computes CDF and simulations of the Fréchet upper bound copula.
 #'
 #' @details
 #'  The Fréchet upper bound copula has CDF :