+++ R/independentCopula.R
@@ -1,7 +1,7 @@
 #' Independence Copula
 #'
 #' @description
-#' Computes CDF of the independence copula.
+#' Computes CDF and simulations of the independence copula.
 #'
 #' @details
 #'  The independence copula has CDF :