+++ man-roxygen/input-template.R
@@ -1,0 +1,1 @@
+#' @param input input for server side.