+++ man-roxygen/output-template.R
@@ -1,0 +1,1 @@
+#' @param output output for server side.